Refactor the CSS to components Refactored the CSS to individual components so it is easier to reason about them.
