+++ components/Footer.tsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import YelpIcon from '../public/YelpIcon';
+import React, { Fragment } from "react";
+import YelpIcon from "../extras/YelpIcon";
 
 /**
  * Used as a simple re-usable component for a service listing on the main page.
@@ -7,18 +7,71 @@ import YelpIcon from '../public/YelpIcon';
 export default class Footer extends React.Component {
   render() {
     return (
-      <footer className="footer">
-        <span>Services provided within 30 miles in any direction of Wilsonville, OR</span>
-        <br/>
-        <div className="socialIcons">
-          <a className="socialIcon" href="https://www.facebook.com/Carpet-Geeks-115408421867268/" target="blank">
-            <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M22.675 0h-21.35c-.732 0-1.325.593-1.325 1.325v21.351c0 .731.593 1.324 1.325 1.324h11.495v-9.294h-3.128v-3.622h3.128v-2.671c0-3.1 1.893-4.788 4.659-4.788 1.325 0 2.463.099 2.795.143v3.24l-1.918.001c-1.504 0-1.795.715-1.795 1.763v2.313h3.587l-.467 3.622h-3.12v9.293h6.116c.73 0 1.323-.593 1.323-1.325v-21.35c0-.732-.593-1.325-1.325-1.325z"/></svg>
-          </a>
-          <YelpIcon width="24px" fill="orange" className="socialIcon" href="https://www.yelp.com/biz/carpet-geeks-wilsonville"/>
-        </div>
-        <br/>
-        <span className="copyrightLine">Copyright Carpet Geeks 2019. Built by <a href="https://tonyneuhold.com/" target="blank">Anton Neuhold</a></span>
-      </footer>
+      <Fragment>
+        <footer className="footer">
+          <span>
+            Services provided within 30 miles in any direction of Wilsonville,
+            OR
+          </span>
+          <br />
+          <div className="socialIcons">
+            <a
+              className="socialIcon"
+              href="https://www.facebook.com/Carpet-Geeks-115408421867268/"
+              target="blank"
+            >
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                width="24"
+                height="24"
+                viewBox="0 0 24 24"
+              >
+                <path d="M22.675 0h-21.35c-.732 0-1.325.593-1.325 1.325v21.351c0 .731.593 1.324 1.325 1.324h11.495v-9.294h-3.128v-3.622h3.128v-2.671c0-3.1 1.893-4.788 4.659-4.788 1.325 0 2.463.099 2.795.143v3.24l-1.918.001c-1.504 0-1.795.715-1.795 1.763v2.313h3.587l-.467 3.622h-3.12v9.293h6.116c.73 0 1.323-.593 1.323-1.325v-21.35c0-.732-.593-1.325-1.325-1.325z" />
+              </svg>
+            </a>
+            <YelpIcon
+              width="24px"
+              fill="orange"
+              className="socialIcon"
+              href="https://www.yelp.com/biz/carpet-geeks-wilsonville"
+            />
+          </div>
+          <br />
+          <span className="copyrightLine">
+            Copyright Carpet Geeks 2019. Built by{" "}
+            <a href="https://tonyneuhold.com/" target="blank">
+              Anton Neuhold
+            </a>
+          </span>
+        </footer>
+
+        <style jsx>{`
+          /* Footer Styling */
+          .footer {
+            margin: 2rem;
+          }
+          .socialIcons {
+            display: flex;
+            flex-direction: row;
+            align-items: center;
+            justify-content: center;
+            margin: 1rem;
+          }
+          :global(.socialIcon) {
+            padding: 1rem;
+            border-radius: 50%;
+            text-align: center;
+            fill: var(--primary-color-orange);
+            display: inline-block;
+          }
+          :global(.socialIcon:hover) {
+            background: var(--primary-color-light-gray);
+          }
+          .copyrightLine {
+            font-size: small;
+          }
+        `}</style>
+      </Fragment>
     );
   }
 }
