+++ components/History.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from "react";
 
 /**
  * Used as a simple re-usable component for a service listing on the main page.
@@ -9,7 +9,22 @@ export default class History extends React.Component {
       <div className="history">
         <h3>Some History...</h3>
         <p>Established in 2011.</p>
-        <p>The Owner Michael Perryman, Started working for a Premier Carpet Cleaning Company in 1997. He Relized the Potential for a Premuim Carpet Cleaning Company, when the outfit he was working for was sold to a New Owner. Michael Started Business in 1999 as a sole-proprietor and had great sucess as an Indepenant Owner/Operator. As the market and technology grew so did the need for his business. Michael evolved his business into a one stop Premium Cleaning & Restoration Company. In 2011, Carpet Geeks was born.</p>
+        <p>
+          The Owner Michael Perryman, Started working for a Premier Carpet
+          Cleaning Company in 1997. He Relized the Potential for a Premuim
+          Carpet Cleaning Company, when the outfit he was working for was sold
+          to a New Owner. Michael Started Business in 1999 as a sole-proprietor
+          and had great sucess as an Indepenant Owner/Operator. As the market
+          and technology grew so did the need for his business. Michael evolved
+          his business into a one stop Premium Cleaning & Restoration Company.
+          In 2011, Carpet Geeks was born.
+        </p>
+        <style jsx>{`
+          .history {
+            max-width: 90vw;
+            margin: 0 auto;
+          }
+        `}</style>
       </div>
     );
   }
