+++ components/Testimonial.tsx
@@ -14,7 +14,7 @@ export default class Testimonial extends React.Component<TestimonialProps> {
           <p className="quote">{testimonial.content}</p>
 
           <h3>
-            <i className="fas fa-quote-right"></i>-{testimonial.userName}
+            <i className="fas fa-quote-right"></i> - {testimonial.userName}
           </h3>
         </div>
         <style jsx>{`
@@ -29,8 +29,6 @@ export default class Testimonial extends React.Component<TestimonialProps> {
             border: 1px solid black;
             box-shadow: 0 1rem 2rem rgba(0, 0, 0, 0.19),
               0 6px 6px rgba(0, 0, 0, 0.23);
-          }
-          .quote {
           }
         `}</style>
       </React.Fragment>
